% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.simple.R
\name{analyze.simple}
\alias{analyze.simple}
\title{analyze.rwty, the main interface for rwty analyses and plots.}
\usage{
analyze.simple(chains, burnin = 0, window.num = 50,
  treespace.points = 100, min.freq = 0, labels = NA,
  likelihood.param = NA, filename = NA, overwrite = FALSE, facet = TRUE,
  ...)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin.  Default value is zero.}

\item{treespace.points}{The number of trees to plot in the treespace plot. Default is 100}

\item{min.freq}{The minimum frequency for a node to be used for calculating discordance. Default is zero.}

\item{filename}{Name of an output file (e.g., "output.pdf").  If none is supplied, rwty will not print outputs to file.}

\item{overwrite}{Boolean variable saying whether output file should be overwritten, if it exists.}

\item{window.size}{The length of window (in trees) for the sliding window plot.  If no value is provided, RWTY selects a number so that 20 windows are analyzed over the chain.}

\item{gens.per.tree}{The number of generations per tree in the .t file.  If no value is provided, RWTY will attempt to determine the number of generations from the tree names.}
}
\value{
output The output is a list containing the following plots:

Plots of likelihood and model parameters as a function of chain length (when p table is available).

Point and heatmap depictions of chains in treespace (points.plot and heatmap).

Sliding window posterior probability, depicting posterior probabilities of the most variable clades
over a series of non-overlapping windows along the MCMC chain.

Cumulative posterior probability, depicting the posterior probabilities of the most variable clades
as a function of chain length.

Sliding window variance, depicting the distribution of variances of clade posterior probability estimates
over a series of non-overlapping windows along the MCMC chain.

Cumulative variance, depicting the distribution of variances in posterior probability estimates as a function of
chain length.

Compare plot (when multiple rwty.trees objects are passed): an xy plot (or several) of the correlation between posterior
estimates from multiple chains.

Discordance tree (when multiple rwty.trees objects are passed): a tree that depicts the similarity in posterior support
values obtained by multiple chains.  Chains that produce similar posterior estimates will appear grouped, while those that
support very different topologies (or posterior support values) will be more distant from each other.
}
\description{
This is the main user interface to rwty.  It allows users to conduct simple
visualizations of MCMC chain performance with  very few arguments.
}
\examples{
data(fungus)
p <- analyze.simple(list(run1, run2), burnin = 50, window.num = 50)
p
}
\keyword{keywords}

