% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.chains.R
\name{check.chains}
\alias{check.chains}
\title{Function for checking suitability of chains for rwty analyses, auto-generating labels, etc}
\usage{
check.chains(chains, labels = NA)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{labels}{The name to use on plots and in generating output files.  If none are provided, defaults are created using Chain 1, Chain 2, etc.}
}
\value{
chains A list of rwty.trees objects
}
\description{
This function is automatically called by many other functions, but can be run manually as well.
It performs a number of tests of chain format, labels, lengths, etc.
}
\examples{
data(fungus)
check.chains(fungus)
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{rwty}

