% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.ptables.R
\name{combine.ptables}
\alias{combine.ptables}
\title{Function for merging p tables for multiple MCMC chains}
\usage{
combine.ptables(chains, burnin = NA)
}
\arguments{
\item{chains}{A list of rwty.chain objects.}

\item{burnin}{The number of trees to omit as burnin. The default (NA) is to use the maximum burnin from all burnins calculated automatically when loading the chains. This can be overidden by providing any integer value.}
}
\value{
ptable A data frame of likelihood values and model parameters for the supplied rwty.chain objects
}
\description{
This function is automatically called by some of the plot functions.
}
\examples{
\dontrun{
data(fungus)
combine.ptables(fungus, burnin=20)
}
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{rwty}
