\name{compare.n}
\alias{compare.n}
\title{A one sentence description of what your function does}
\usage{
  compare.n(x, setnames = NA, burnin)
}
\arguments{
  \item{inputParameter1}{A description of the input
  parameter \code{inputParameter1}}

  \item{inputParameter2}{A description of the input
  parameter \code{inputParameter2}}
}
\value{
  output A description of the object the function outputs
}
\description{
  A more detailed description of what the function is and
  how it works. It may be a paragraph that should not be
  separated by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

