% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwty.utility.functions.R
\name{continuous.distance}
\alias{continuous.distance}
\title{Various small utility functions}
\usage{
continuous.distance(two.trees)
}
\description{
A bunch of tiny utility functions that we need off and on.  Not intended for user interaction.
}

