% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discordance.mat.R
\name{discordance.mat}
\alias{discordance.mat}
\title{Calculate discordance metric}
\usage{
discordance.mat(freqTable, min.freq = 0)
}
\arguments{
\item{min.freq}{The minimum frequency for a node to be used for calculating discordance}

\item{x}{A table of clade frequencies.}
}
\value{
output A dist class object of discordance values
}
\description{
Calculate discordance metric
}
\examples{
discordance.mat(myfreqTable)
}
\keyword{MCMC,}
\keyword{discordance}
\keyword{phylogenetics,}

