% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discordance.n.R
\name{discordance.n}
\alias{discordance.n}
\title{Perform sliding window analysis of discordance among chains.}
\usage{
discordance.n(x, setnames = NA, min.freq = 0)
}
\arguments{
\item{x}{A list generated by analyze.multi}

\item{setnames}{A list of names for the chains.}

\item{min.freq}{The minimum frequency for a node to be used for calculating discordance}
}
\value{
output A list containing for each window: a table of frequencies of each clade in each chain, a distance matrix measuring discordance, and an hclust plot.
}
\description{
This function takes analyze.multi output and an optional array of names for those
files.  It returns a plot of the change in discordance
}
\examples{
discordance.n(myAnalyze.multi.object, setnames=c("Chain 1", "Chain 2", "Chain 3"))
}
\keyword{MCMC,}
\keyword{discordance,}
\keyword{phylogenetics,}
\keyword{sliding}
\keyword{window}

