% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.autocorr.k.R
\name{estimate.autocorr.k}
\alias{estimate.autocorr.k}
\title{Calculate sampling interval based on exponential semivariogram model.}
\usage{
estimate.autocorr.k(dat, ac.cutoff = 0.95)
}
\arguments{
\item{dat}{A data frame output from topological.autocorr.}

\item{ac.cutoff}{The proportion of the asymptotic path distance to use as a cutoff for determining sampling interval.}
}
\value{
A data frame consisting of the k value matching the ac.cutoff proportion of the asymptotic 
path distance for each chain.  This sampling interval estimates the interval at which path distances
are no longer autocorrelated.
}
\description{
This function uses an exponential semivariogram model to estimate the asymptotic path distance, and
uses that to estimate the sampling interval at which path distances have reached some fixed
proportion of that value (default 0.95).  It expects as input a data table output by rwty's
}
\examples{
data(fungus)
sampling.table <- topological.autocorr(fungus, burnin = 20, max.intervals = 40)
estimate.autocorr.k(sampling.table, ac.cutoff = 0.9)
}
\keyword{autocorrelation,}
\keyword{distance}
\keyword{path}

