% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.multi.R
\name{load.multi}
\alias{load.multi}
\title{Load all matching files from a directory into a list of rwty.trees objects}
\usage{
load.multi(path = ".", ext.tree = "t", ext.p = "p", labels = NA, ...)
}
\arguments{
\item{path}{The path to the directory containing tree and log files}

\item{ext.tree}{The extension to be used in finding tree files.  Usually should be "t" or "trees", or something of that sort.}

\item{ext.p}{The extension to be used in finding log files.
Basically the function is going to chop ext.tree off of each file name and glue ext.p onto it, so "myfile.t" will look for a logfile "myfile.p", if ext.tree = "t" and ext.p = "p".}
}
\value{
output A list of rwty.trees objects containing the multiPhylos and the tables of values from the log files if available.
}
\description{
Loads trees, looks for a .p file of tree likelihoods, returns rwty.trees objects containing both
}
\examples{
#load.multi(path = "~/my trees/", ext.tree = "t", ext.p = "log")
}
\keyword{MCMC,}
\keyword{Phylogenetics,}
\keyword{load,}
\keyword{trees}

