% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.trees.R
\name{load.trees}
\alias{load.trees}
\title{Custom functions to load tree lists so that rwty can do basic processing on the way in.}
\usage{
load.trees(file, type = NA, format = "mb", gens.per.tree = NA,
  trim = "auto", logfile = NA, skip = NA, comment.chars = c("[",
  "#"), treedist = "RF", burnin = NA, lnl.column = NA)
}
\arguments{
\item{file}{A path to a tree file containing an MCMC chain of trees}

\item{type}{An argument that designates the type of tree file.  If "nexus",
trees are loaded using ape's \code{\link{read.nexus}} function.  Otherwise, it's \code{\link{read.tree}}.
If a "format" argument is passed, type will be determined from the format definition.}

\item{format}{File format, which is used to find tree and log files, and to rename the likelihood column.
Currently accepted values are "mb" for MrBayes, "beast" for BEAST, "*beast" for *BEAST, and "revbayes" for RevBayes.
If you would like RWTY to understand additional formats, please contact the authors and send us some sample data.}

\item{gens.per.tree}{The number of generations separating trees.  If not provided, RWTY will attempt to calculate it automatically.}

\item{trim}{Used for thinning the chain.  If a number N is provided, RWTY keeps every Nth tree.  If trim is set to "auto" or TRUE, RWTY automatically selects a trim value and subsamples the chain so that it has around 1000 trees.  If trim is set to "TRUE" or 1, all trees are retained.}

\item{logfile}{A path to a file containing model parameters and likelihoods.  If no path is provided but
a "format" argument is supplied, RWTY will attempt to find the log file automatically based on the format
definition.  RWTY will then use the "format" argument to find the likelihood column and rename it to "LnL".}

\item{skip}{The number of lines that must be skipped to get to the header of the log file.
MrBayes, for instance, prints a comment line at the top of the log file, so MrBayes files should be
read in with a skip value of 1.  If no "skip" value is provided but a "format" is supplied, RWTY will
attempt to read logs using the skip value from the format definition.}

\item{comment.chars}{A vector of characters that indicate that a line in a log file is a comment and should not be interpreted.  Will be used to attempt to automatically determine an approporiate "skip" value if none is provided.}

\item{treedist}{the type of tree distance metric to use, can be 'PD' for path distance or 'RF' (the default) for Robinson Foulds distance}

\item{burnin}{the number of samples at the start of the chain to exclude as burnin. The default (burnin = NA) is to calculate the burnin automatically if there is a logfile with likelihood values, or to assume that the burnin is 25\% if there is no logfile with likelihood values.}

\item{lnl.column}{the name of the likelihood column in the log file(s).  Only needs to be specified when it differs from the name chosen automatically from the "format" argument.}
}
\value{
output An rwty.chain object containing the multiPhylo and the table of values from the log file if available.
}
\description{
Loads trees, looks for a log file of tree likelihoods and parameter values, returns an rwty.chain object containing both
}
\examples{
#load.trees(file="mytrees.t", format = "mb")
}
\seealso{
\code{\link{read.tree}}, \code{\link{read.nexus}}
}
\keyword{MCMC,}
\keyword{Phylogenetics,}
\keyword{load}
