% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeanim.treespace.R
\name{makeanim.treespace}
\alias{makeanim.treespace}
\title{Plot chains in treespace.}
\usage{
makeanim.treespace(chains, burnin = NA, n.points = 200, fill.color = "LnL")
}
\arguments{
\item{chains}{A list of one or more rwty.chain objects}

\item{burnin}{The number of trees to omit as burnin. The default (NA) is to use the maximum burnin from all burnins calculated automatically when loading the chains. This can be overidden by providing any integer value.}

\item{n.points}{The minimum number of points on each plot. The function will automatically choose the thinning value which gets you the smallest number of trees that is at least as much as this value. The default (200) is usually sufficient to get a good idea of what is happening in your chains.}

\item{fill.color}{The name of the column from the log table that that you would like to use to colour the points in the plot. The default is to colour the points by LnL.}
}
\value{
A gganimate object which has one frame for each of the sampled generations.
}
\description{
This function will take list of rwty.chains objects and produce animations of chains in treespace.
}
\examples{
\dontrun{
data(fungus)

# Make the animation (can take a few minutes)
a <- makeanim.treespace(salamanders[1,2])


# Save the animation
# see: https://gganimate.com/reference/anim_save.html
# Note that the width and height of the animation can be fiddly to set, 
# so using this function as below will save the last animation, in which the with and height
# have (hopefully) been set sensibly by default.
# By default RWTY produces GIFs, but you can change this using the function below
# and using the information from the gganimate link above.
anim_save("~/Desktop/animation.gif")
}
}
\keyword{animation}
\keyword{gif}
\keyword{plot,}
\keyword{rwty,}
\keyword{treespace,}
