% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplot.asdsf.mb.R
\name{makeplot.asdsf.mb}
\alias{makeplot.asdsf.mb}
\title{Plot the Standard Deviation of Split Frequencies over the course of an MCMC
with sliding windows growing in size in the style of MrBayes.}
\usage{
makeplot.asdsf.mb(
  chains,
  burnin = 0,
  window.lookback = 0.75,
  window.number = 10,
  min.freq = 0,
  log.y = TRUE
)
}
\arguments{
\item{chains}{A list of rwty.chain objects.}

\item{burnin}{The number of trees to eliminate as burnin. Defaults to zero.}

\item{window.lookback}{The percentage of preceding samples to use for each estimate.}

\item{window.number}{The number of ASDSFs estimates to plot at evenly spaced number of iterations after burnin and including the final sample.}

\item{min.freq}{The minimum frequency for a node to be used for calculating ASDSF.}

\item{log.y}{Controls whether they Y axis is plotted on a log scale or not.  Which scale is more useful depends largely on the amount of disagreement between your chains.  Attempting to make an asdsf plot with a log Y axis for chains that include standard deviations of zero will result in warning messages.}
}
\value{
output A cumulative plot of ASDSF across all chains
}
\description{
This function takes two or more rwty.chain ojects and returns a plot of
ASDSF at number.points evenly spaced iterations.
Each ASDSF is calculated on the most recent window.lookback of samples
The solid line with points shows the Average Standard Deviation of Split Frequences at the current generation
}
\examples{
\dontrun{
    data(fungus)
    p <- makeplot.asdsf.mb(fungus, burnin = 20)
    p
}
}
\keyword{ASDSF,}
\keyword{MCMC,}
\keyword{cumulative}
\keyword{phylogenetics,}
