% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.asdsf.R
\name{makeplot.asdsf}
\alias{makeplot.asdsf}
\title{Plot the Average Standard Deviation of Split Frequencies in sliding windows.}
\usage{
makeplot.asdsf(chains, burnin, window.num, min.freq = 0.1, ...)
}
\arguments{
\item{chains}{A list of two or more rwty.trees objects}

\item{burnin}{The number of samples to remove from the start of the chain as burnin}

\item{window.num}{The number of windows to use for generating plots}

\item{min.freq}{The minimum frequency for a node to be used for calculating ASDSF}
}
\value{
output A sliding window plot of ASDSF across all chains
}
\description{
This function takes two or more rwty.trees ojects and returns a sliding window plot of ASDSF
}
\examples{
data(fungus)
test.chains <- list(run1, run2)
test.chains <- check.chains(test.chains)
p <- makeplot.asdsf(chains = test.chains, burnin = 100, window.num = 20)
p
}
\keyword{ASDSF,}
\keyword{MCMC,}
\keyword{phylogenetics,}
\keyword{sliding}
\keyword{window}

