% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplot.autocorr.R
\name{makeplot.autocorr}
\alias{makeplot.autocorr}
\title{Make autocorrelation plots of tree topologies from MCMC analyses}
\usage{
makeplot.autocorr(chains, burnin = 0, autocorr.intervals = 100,
  squared = FALSE, facet = FALSE, free_y = FALSE, treedist = "PD")
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin.}

\item{autocorr.intervals}{The maximum number of sampling intervals to use.}

\item{squared}{TRUE/FALSE use squared tree distances (necessary to calculate approximate ESS; default FALSE)}

\item{facet}{TRUE/FALSE to turn facetting of the plot on or off (default FALSE)}

\item{free_y}{TRUE/FALSE to turn free y scales on the facetted plots on or off (default FALSE). Only works if facet = TRUE.}

\item{treedist}{the type of tree distance metric to use, can be 'PD' for path distance or 'RF' for Robinson Foulds distance}
}
\value{
A ggplot2 plot object, with one line (facetting off) or facet
(facetting on) per rwty.trees object.
}
\description{
This function takes a list of rwty.trees objects, and makes an
autocorrelation plot for each chain. Each plot shows the median phylogenetic
distance at a series of sampling intervals. In really well behaved MCMC 
analyses, the median distance will stay constant as the sampling interval 
increases. If there is autocorrelation, the median distance will 
increase as the sampling interval increases, and is expected to level
off when the autocorrelation decreases to zero. The function calculates
path distances, though other distances could also be employed.
}
\examples{
data(fungus)
makeplot.autocorr(fungus, burnin = 20)
}
\keyword{autocorrelation,}
\keyword{distance}
\keyword{path}

