% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.cladeprobs.sliding.R
\name{makeplot.cladeprobs.sliding}
\alias{makeplot.cladeprobs.sliding}
\title{Plot clade probabilities in sliding windows over the course of an MCMC}
\usage{
makeplot.cladeprobs.sliding(chains, burnin = 0, n.clades = 20,
  window.size = 20, facet = TRUE)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}

\item{n.clades}{The number of clades to plot}

\item{window.size}{The number of trees to include in each window (note, specified as a number of sampled trees, not a number of generations)}

\item{facet}{(TRUE/FALSE). TRUE: return a single plot with one facet per chain; FALSE: return a list of individual plots with one plot per chain}
}
\value{
cladeprobs.plot Either a single ggplot2 object or a list of ggplot2 objects.
}
\description{
Takes a list of rwty.trees objects.
Plots the posterior probabilities of clades over the course of the MCMC, calculated from windows of a specified size.
Only plots the n.clades most variable clades, as measured by the standard deviation of the posterior probabilities of each clade across all windows.
Each line in the plot represents a single clade. The colour of the line represents the standard deviation of the posterior probabilities of that clade across all sliding windows.
}
\examples{
data(fungus)
makeplot.cladeprobs.sliding(fungus, burnin = 20, n.clades=25)
}
\keyword{mcmc,}
\keyword{phylogenetics,}
\keyword{plot}
\keyword{sliding}
\keyword{window,}

