% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.cladevar.R
\name{makeplot.cladevar}
\alias{makeplot.cladevar}
\title{Plots distribution of variance in posterior probability estimates across clades.}
\usage{
makeplot.cladevar(input.table, numclades = nrow(input.table))
}
\arguments{
\item{input.table}{Either a rwty.cumulative or rwty.slide object, or a cumulative or slide table from one of those objects.}

\item{numclades}{The number of clades to calculate stats for.  Defaults to all clades in the table.}
}
\value{
thisplot A ggplot object
}
\description{
Plots the distribution of variances in posterior probability estimates
across clades in a chain as a function of chain length.
}
\examples{
data(fungus)
slide.data <- slide.freq(fungus$Fungus.Run1$trees, burnin=100, window.size=20, gens.per.tree=10000)
makeplot.cladevar(slide.data$slide.table, numclades=100)
}
\keyword{convergence,}
\keyword{mcmc,}
\keyword{phylogenetics,}
\keyword{uncertainty}

