% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.cumulative.asdsf.R
\name{makeplot.cumulative.asdsf}
\alias{makeplot.cumulative.asdsf}
\title{Plot the variance and cumulative Average Standard Deviation of Split Frequencies.}
\usage{
makeplot.cumulative.asdsf(chains, burnin, window.num, min.freq = 0.1, ...)
}
\arguments{
\item{chains}{A list of two or more rwty.trees objects}

\item{burnin}{The number of samples to remove from the start of the chain as burnin}

\item{window.num}{The number of windows to use for generating plots}

\item{min.freq}{The minimum frequency for a node to be used for calculating ASDSF}
}
\value{
output A cumulative plot of ASDSF across all chains
}
\description{
This function takes two or more rwty.trees ojects and returns a cumulative plot of ASDSF
}
\examples{
data(fungus)
test.chains <- check.chains(fungus)
p <- makeplot.asdsf(chains = test.chains, burnin = 100, window.num = 20)
p
}
\keyword{ASDSF,}
\keyword{MCMC,}
\keyword{cumulative}
\keyword{phylogenetics,}

