% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplot.cumulative.tree.R
\name{makeplot.cumulative.tree}
\alias{makeplot.cumulative.tree}
\title{Plot convergence metrics for nodes on a consensus tree}
\usage{
makeplot.cumulative.tree(chains, burnin = 0, window.size = 20,
  rank = "wcsf", p = 0.5)
}
\arguments{
\item{chains}{A list of rwty.chain objects.}

\item{burnin}{The number of trees to eliminate as burnin}

\item{window.size}{The number of trees to include in each window (note, specified as a number of sampled trees, not a number of generations)}

\item{rank}{('wcsf', 'sd'). How to rank the clades? By default, we plot the 20 'worst' clades. This parameter sets the definition of 'worst'. The default is to rank the by the weighted change in split frequencies (rank = 'wcsf'). This works by looking at the change in the cumulative split frequency over the course of the MCMC, and ranks the worst chains as those that do not level off (i.e. those that have changes near the end). We do this because in a well-behaved chain, we expect the cumulative split frequencies to level off once the chain has been run for long enough. So, any cumulative split frequencies which are still changing towards the end of your run are likely to indicate problematic clades. Specifically, we multiply the absolute change in split frequencies for each clade by a set of weights that increase linearly towards the end of the chain (the first observation gets a weight of zero, the final observation gets a weight of one). The original AWTY ranked clades by their standard deviations (higher SD = worse), so we include this as an option too. To do this, just set rank = 'sd'.}

\item{p}{The minimum frequency of a bipartition to appear in the consensus tree.}
}
\value{
cumulative.tree A consensus tree for each chain with node values representing the wcsf or standard deviation for that node.
}
\description{
Takes a list of rwty.chain objects.  
Plots the standard deviation or wcsf for nodes in the consensus tree for each chain.  These values represent the consistency of the posterior probability estimate for that node along the chain.
The function returns one tree per chain.
}
\examples{
data(fungus)
makeplot.cumulative.tree(fungus, burnin = 20, window.size = 20)
}
\keyword{convergence,}
\keyword{mcmc,}
\keyword{phylogenetics,}
\keyword{plot}
