% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.posteriors.R
\name{makeplot.posteriors}
\alias{makeplot.posteriors}
\title{Plot posterior probabilities of clades over an MCMC chain}
\usage{
makeplot.posteriors(chains, burnin, window.num, ...)
}
\arguments{
\item{chains}{A list of one or more rwty.trees objects}

\item{burnin}{The number of samples to remove from the start of the chain as burnin}

\item{window.num}{The number of windows to use for generating plots}

\item{...}{Extra arguments to be passed to plotting functions.}
}
\value{
A list of cumulative and sliding window plots
}
\description{
This function will take a set of chains and produce plots of posteriors
}
\examples{
data(fungus)
makeplot.posteriors(fungus, burnin = 100, window.num = 20)

# NB: these data indicate significant problems: the two chains are sampling very different parts of tree space
}
\keyword{plot,}
\keyword{posterior}
\keyword{probabilities,}
\keyword{rwty}

