% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.splitfreq.matrix.R
\name{makeplot.splitfreq.matrix}
\alias{makeplot.splitfreq.matrix}
\title{Plots a matrix of split frequency comparisons between multiple MCMC chains.}
\usage{
makeplot.splitfreq.matrix(chains, burnin = 0)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}
}
\value{
output A list containing a table of frequencies of each clade in each chain along with mean and sd, a distance matrix measuring consensus between chains, a translation table, and a ggpairs plot.
}
\description{
This function takes list of rwty.trees objects, and returns a scatterplot matrix
in which each plot shows the split frequencies of all clades that appear in one or both
MCMC chains at least once. In the upper diagonal, we show the correlation between the split
frequencies (Pearson's R), and the Average Standard Deviation of the split frequencies.
}
\examples{
data(fungus)
makeplot.splitfreq.matrix(salamanders[1:4], burnin = 20)
}
\keyword{MCMC,}
\keyword{clade}
\keyword{consensus,}
\keyword{convergence}
\keyword{frequency,}
\keyword{phylogenetics,}

