% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplot.cladeprobs.cumulative.R
\name{makeplot.splitfreqs.cumulative}
\alias{makeplot.splitfreqs.cumulative}
\title{Plot cumulative split frequencies over the course of an MCMC}
\usage{
makeplot.splitfreqs.cumulative(chains, burnin = 0, n.clades = 20,
  window.size = 20, facet = TRUE)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}

\item{n.clades}{The number of clades to plot}

\item{window.size}{The number of trees to include in each window (note, specified as a number of sampled trees, not a number of generations)}

\item{facet}{(TRUE/FALSE). TRUE: return a single plot with one facet per chain; FALSE: return a list of individual plots with one plot per chain}
}
\value{
splitfreqs.plot Either a single ggplot2 object or a list of ggplot2 objects.
}
\description{
Takes a list of rwty.trees objects.  
Plots the cumulative split frequencies of clades over the course of the MCMC. Stationarity is indicated by split frequencies levelling out.
Only plots the n.clades most variable clades, as measured by the standard deviation of the split frequencies of each clade across all windows.  
Each line in the plot represents a single clade. The colour of the line represents the standard deviation of the split frequencies of that clade across all sliding windows.
}
\examples{
data(fungus)
makeplot.splitfreqs.cumulative(fungus, burnin = 20, n.clades=25)
}
\keyword{mcmc,}
\keyword{phylogenetics,}
\keyword{plot}
\keyword{sliding}
\keyword{window,}

