% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.topology.R
\name{makeplot.topology}
\alias{makeplot.topology}
\title{Plotting parameters}
\usage{
makeplot.topology(chains, burnin = 0, facet = TRUE, free_y = TRUE)
}
\arguments{
\item{chains}{A set of rwty.trees objects.}

\item{burnin}{The number of trees to omit as burnin.}

\item{facet}{Boolean denoting whether to make a facet plot.}

\item{free_y}{TRUE/FALSE to turn free y scales on the facetted plots on or off (default TRUE). Only works if facet = TRUE.}
}
\value{
topology.plot Returns a ggplot object.
}
\description{
Plots tree topologies over the length of the MCMC chain. The plot shows the path distance
of each tree in the chain from the first tree (after discarding burnin)
}
\examples{
data(fungus)
makeplot.topology(fungus, burnin=20, parameter="pi.A.")
}
\keyword{convergence,}
\keyword{mcmc,}
\keyword{parameter,}
\keyword{phylogenetics}
\keyword{plot,}

