% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.topology.trace.R
\name{makeplot.topology.trace}
\alias{makeplot.topology.trace}
\title{Plotting parameters}
\usage{
makeplot.topology.trace(chains, burnin = 0, facet = TRUE, free_y = FALSE,
  independent.chains = FALSE)
}
\arguments{
\item{chains}{A set of rwty.trees objects.}

\item{burnin}{The number of trees to omit as burnin.}

\item{facet}{TRUE/FALSE denoting whether to make a facet plot (default TRUE)}

\item{free_y}{TRUE/FALSE to turn free y scales on the facetted plots on or off (default FALSE). Only works if facet = TRUE.}

\item{independent.chains}{TRUE/FALSE if FALSE (the default) then the plots show the distance of each tree from the first tree of the first chain. If TRUE, the plots show the distance of each tree from the first tree of the chain in which that tree appears. The TRUE option should only be used in the case that different chains represent analyses of different genes or datasets.}
}
\value{
topology.plot Returns a ggplot object.
}
\description{
Plots tree topologies over the length of the MCMC chain. The plot shows the path distance
of each tree in each chain from the first tree of the first chain (after discarding burnin).
If required, the behaviour can be changed to plot the path distance of each tree from the first
tree in each chain, using the independent.chains option.
}
\examples{
data(fungus)
makeplot.topology.trace(fungus, burnin=20, parameter="pi.A.")
}
\keyword{convergence,}
\keyword{mcmc,}
\keyword{parameter,}
\keyword{phylogenetics}
\keyword{plot,}

