% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.treespace}
\alias{plot.treespace}
\title{Plot chains in treespace.}
\usage{
\method{plot}{treespace}(points)
}
\arguments{
\item{points}{A set of points to plot \code{points}}
}
\value{
p A ggplot object
}
\description{
This function will take an mds.treespace object and produce plots of chains in treespace.
}
\keyword{plot,}
\keyword{rwty}
\keyword{treespace,}

