% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rwty.chain.R
\name{print.rwty.chain}
\alias{print.rwty.chain}
\title{Function for printing rwty.chain objects}
\usage{
\method{print}{rwty.chain}(x, ...)
}
\arguments{
\item{x}{A rwty.chain object}

\item{...}{Other arguments to be passed on to next function}
}
\value{
A summary of the contents of the chain
}
\description{
This function is automatically called when printing a rwty.chain object
}
\examples{
data(fungus)
fungus$Fungus.Run1
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{plot,}
\keyword{rwty}

