% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.rwty.trees.R
\name{print.rwty.trees}
\alias{print.rwty.trees}
\title{Function for printing rwty.trees objects}
\usage{
\method{print}{rwty.trees}(x)
}
\arguments{
\item{x}{A rwty.trees object}
}
\value{
A summary of the contents of the chain
}
\description{
This function is automatically called when printing a rwty.trees object
}
\examples{
data(fungus)
run1
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{plot,}
\keyword{rwty}

