% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Start the RWTY Shiny web app}
\usage{
run_shiny()
}
\arguments{
\item{...}{arguments passed to \code{\link[shiny]{runApp}}}
}
\description{
Runs the RWTY Shiny web app, for interactive visualization of convergence.
}
\examples{
\dontrun{
run_shiny(...)
}
}

