\name{rwty-package}
\alias{rwty-package}
\title{RWTY: A package for visualizing chains in MCMC Bayesian phylogenetics}
\description{
  This package implements various tests, visualizations, and metrics
    for diagnosing convergence of MCMC chains in phylogenetics.  It implements
    and automates many of the functions of the AWTY package in the R
}
