% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topological.autocorr.R
\name{topological.autocorr}
\alias{topological.autocorr}
\title{Calculate data for autocorrelation plots of tree topologies from MCMC analyses}
\usage{
topological.autocorr(chains, burnin = 0, max.intervals = 100)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}

\item{max.intervals}{The maximum number of sampling intervals to use}
}
\value{
A data frame with one row per sampling interval, per chain.
The first column is the sampling interval. The second column is the median
path distance between pairs of trees from that sampling interval. The third
column is the chain ID.
}
\description{
This function takes a list of rwty.trees objects, and calculates the
median phylogenetic distance at a series of roughly even sampling intervals.
In really well behaved MCMC analyses, the median distance will stay constant
as the sampling interval increases. If there is autocorrelation, it will
increase as the sampling interval increases, and is expected to level
off when the autocorrelation decreases to zero. The function calculates
path distances, though other distances could also be employed.
}
\examples{
data(fungus)
topological.autocorr(fungus, burnin = 100)
}
\keyword{autocorrelation,}
\keyword{distance}
\keyword{path}

