% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tree.distance}
\alias{tree.distance}
\title{Returns distance between two trees}
\usage{
tree.distance(two.trees)
}
\arguments{
\item{two.trees}{A list containing two trees \code{two.trees}}
}
\value{
d A distance measurement
}
\description{
Takes a list of two trees, returns Robinson-Foulds distance
between them.
}
\examples{
data(fungus)
tree.distance(list(run1$trees[[1]], run1$trees[[2]]))
}
\keyword{Foulds}
\keyword{Robinson}
\keyword{distance,}
\keyword{tree}

