% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topological.ess.R
\name{tree.ess}
\alias{tree.ess}
\title{Calculate the approximate Effective Sample Size (ESS) of tree topologies}
\usage{
tree.ess(tree.list)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}

\item{n}{The number of replicate analyses to do}
}
\value{
A data frame with one row per chain, and columns describing the
median ESS, the upper and lower 95% confidence intervals based on the
replicates performed, and the name of the chain.
}
\description{
This function takes a list of rwty.trees objects, and calculates the
approximate ESS of the trees from each chain, after removing burnin.
Each caulcation is repeated n times, where in each replicate a random
tree from the chain is chosen as a 'focal' tree. The calculation works
by calculating the Robinson Foulds distance of each tree in the chain
from the focal tree, and calculating the ESS of the resulting vector
of phylogenetic distances using the effectiveSize function from the
coda package. NB this function requires the calculation of many many
tree distances, so can take some time.
}
\examples{
data(fungus)
topological.ess(chains = list(run1, run2), burnin = 250, n = 10)
}
\keyword{distance,}
\keyword{robinson-foulds}
\keyword{tree}
\keyword{treespace,}

