% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treespace.R
\name{treespace}
\alias{treespace}
\title{MDS scaling of treespace for a single tree list.}
\usage{
treespace(chains, min.points = 200, burnin = NA, fill.color = "LnL")
}
\arguments{
\item{chains}{A list of 1 or more rwty.chain objects.}

\item{min.points}{The minimum number of points on each plot. The function will automatically choose the thinning value which gets you the smallest number of trees that is at least as much as this value. The default (200) is usually sufficient to get a good idea of what is happening in your chains.}

\item{burnin}{The number of trees to omit as burnin. The default (NA) is to use the maximum burnin from all burnins calculated automatically when loading the chains. This can be overidden by providing any integer value.}

\item{fill.color}{The name of the column from the log table that that you would like to use to colour the points in the plot. The default is to colour the points by LnL.}
}
\value{
Returns a list containing the points
}
\description{
This function constructs a distance matrix from a list of trees
and uses multi-dimensional scaling to collapse it to a two-
dimensional tree space for plotting.
}
\examples{
\dontrun{
data(fungus)
treespace(fungus)
}
}
\keyword{mds,}
\keyword{multi-dimensional}
\keyword{scaling}
\keyword{treespace,}
