% Generated by roxygen2 (4.0.1): do not edit by hand
\name{treespace.single}
\alias{treespace.single}
\title{MDS scaling of treespace for a single tree list.}
\usage{
treespace.single(trees, gens, ptable = NULL)
}
\arguments{
\item{trees}{A multiphylo object \code{trees}}

\item{gens}{A list of the generation represented by each tree, for use in plotting \code{}}

\item{ptable}{A table of p values to use for plotting likelihoods \code{p.file}}
}
\value{
Returns a list containing the mds table and a plot.
}
\description{
This function constructs a distance matrix from a list of trees
and uses multi-dimensional scaling to collapse it to a two-
dimensional tree space for plotting.
}
\examples{
data(fungus)
burnin <- 100
#We have to start by trimming the chain and p table down to a reasonable number of trees
mdstrees <- run1$trees[seq((burnin + 1), length(run1$trees), by = 20)]
#Now we're going to get a list of generations matching the trees retained from the chain
gens <- as.numeric(unlist(regmatches(names(run1$trees), gregexpr('\\\\(?[0-9]+', names(run1$trees)))))
gens <- gens[seq((burnin + 1), length(run1$trees), by = 20)]
#Finally we're going to cut down the p table to just the bits we're going to use
mdsptable <- run1$ptable[seq((burnin + 1), length(run1$trees), by = 20),]
this.treespace <- treespace.single(mdstrees, gens=gens, ptable=mdsptable)
}
\keyword{mds,}
\keyword{multi-dimensional}
\keyword{scaling}
\keyword{treespace,}

