% Generated by roxygen2 (4.0.1): do not edit by hand
\name{analyze.multi}
\alias{analyze.multi}
\title{Function for running rwty analyses on multiple chains.}
\usage{
analyze.multi(chains, burnin, window.size, gens.per.tree = NA,
  treespace.points = 100, filename = NA, labels = NA, ...)
}
\arguments{
\item{chains}{A list of rwty.trees objects. \code{chains}}

\item{burnin}{The number of trees to eliminate as burnin \code{burnin}}

\item{window.size}{The length of window (in trees) for the sliding window plot \code{window.size}}

\item{gens.per.tree}{The number of generations per tree in the .t file.}

\item{treespace.points}{The number of trees to plot in the treespace plot.}

\item{filename}{A name to be used for generating pdfs of output.  \code{filename}}

\item{labels}{The name to use on plots and in generating output files.  \code{labels}}
}
\value{
output A list of outputs from the analyze.single runs on each chain, as well as a compare.n run for all chains.  Eventually we will add more multi-chain analyses.
}
\description{
This function is automatically called when analyze.rwty is called with multple chains.
It runs analyze.single for each chain, and then does multi-chain analyses as well.
}
\examples{
data(fungus)
analyze.multi(list(run1, run2), burnin=100, window.size=20, treespace.points=100, filename="fungus.pdf", labels=c("Chain1", "Chain2"))
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{plot,}
\keyword{rwty}

