% Generated by roxygen2 (4.0.1): do not edit by hand
\name{analyze.rwty}
\alias{analyze.rwty}
\title{Analyze.rwty, the main interface for rwty analyses and plots.}
\usage{
analyze.rwty(chains, burnin, window.size, gens.per.tree = NA,
  treespace.points = 100, ...)
}
\arguments{
\item{chains}{A list of rwty.trees objects. \code{chains}}

\item{burnin}{The number of trees to eliminate as burnin \code{burnin}}

\item{window.size}{The length of window (in trees) for the sliding window plot \code{window.size}}

\item{gens.per.tree}{The number of generations per tree in the .t file.}

\item{treespace.points}{The number of trees to plot in the treespace plot.}
}
\value{
output A list of outputs from the analyze.single runs on each chain, as well as a compare.n run for all chains.  Eventually we will add more multi-chain analyses.
}
\description{
This is the main user interface to rwty.  It allows users to chuck in arguments for
chains, burnin, window size, gens per tree, and "step", and returns an object that
contains sliding window and cumulative posterior probability plots, treespace plots,
and multi-chain diagnostic plots when multiple chains are provided.
}
\examples{
data(fungus)
single <- analyze.rwty(run1, burnin=100, window.size=20, treespace.points=50, filename="Run1.pdf")
multi <- analyze.rwty(list(run1, run2), burnin=100, window.size=20, treespace.points=50, labels=c("Chain1", "Chain2"), filename="multi analysis.pdf")
}
\keyword{keywords}

