% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.rwty.R
\name{analyze.rwty}
\alias{analyze.rwty}
\title{analyze.rwty, the main interface for rwty analyses and plots.}
\usage{
analyze.rwty(chains, burnin = 0, window.num = 50, treespace.points = 100,
  min.freq = 0.1, labels = NA, likelihood.param = NA, filename = NA,
  overwrite = FALSE, facet = TRUE, ess.reps = 50,
  autocorr.intervals = 100, ...)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin.  Default value is zero.}

\item{treespace.points}{The number of trees to plot in the treespace plot. Default is 100}

\item{min.freq}{The minimum frequency for a node to be used for calculating ASDSF. Default is 0.1}

\item{filename}{Name of an output file (e.g., "output.pdf").  If none is supplied, rwty will not print outputs to file.}

\item{overwrite}{Boolean variable saying whether output file should be overwritten, if it exists.}

\item{window.size}{The length of window (in trees) for the sliding window plot.  If no value is provided, RWTY selects a number so that 20 windows are analyzed over the chain.}

\item{gens.per.tree}{The number of generations per tree in the .t file.  If no value is provided, RWTY will attempt to determine the number of generations from the tree names.}
}
\value{
output A list of outputs from the analyze.single runs on each chain, as well as a compare.n run for all chains.  Eventually we will add more multi-chain analyses.
}
\description{
This is the main user interface to rwty.  It allows users to chuck in arguments for
chains, burnin, window size, gens per tree, and "step", and returns an object that
contains sliding window and cumulative posterior probability plots, treespace plots,
and multi-chain diagnostic plots when multiple chains are provided.
}
\examples{
data(fungus)
p <- analyze.rwty(fungus, burnin = 50, window.num = 50)
p
}
\keyword{keywords}

