% Generated by roxygen2 (4.0.1): do not edit by hand
\name{analyze.single}
\alias{analyze.single}
\title{Function for running rwty analyses on single chains.}
\usage{
analyze.single(chains, burnin = 0, window.size, gens.per.tree = NA,
  treespace.points = 100, filename = NA, labels = NA, treespace = TRUE,
  ...)
}
\arguments{
\item{chains}{A single rwty.trees object. \code{chains}}

\item{burnin}{The number of trees to eliminate as burnin \code{burnin}}

\item{window.size}{The length of window (in trees) for the sliding window plot \code{window.size}}

\item{gens.per.tree}{The number of generations per tree in the .t file. \code{gens.per.tree}}

\item{treespace.points}{The number of trees to plot in the treespace plot.  \code{treespace.points}}

\item{filename}{A name to be used for generating pdfs of output.  \code{filename}}

\item{labels}{The name to use on plots and in generating output files.  \code{labels}}

\item{treespace}{Boolean to determine whether or not treespace plots are made. \code{treespace}}
}
\value{
output A list with tables and plots for LnL (when .p file exists), sliding window,
cumulative, and treespace plots.
}
\description{
This function is automatically called when analyze.rwty is called with one chain.
It runs analyze.single on the chain and returns an object with the appropriate tables
and plots.
}
\examples{
data(fungus)
analyze.single(run1, burnin=100, window.size=20, treespace.points=100, filename="fungus.pdf", labels="Chain1")
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{plot,}
\keyword{rwty}

