% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.chains.R
\name{check.chains}
\alias{check.chains}
\title{Function for checking suitability of chains for rwty analyses, auto-generating labels, etc}
\usage{
check.chains(chains)
}
\arguments{
\item{chains}{A list of rwty.chain objects.}
}
\value{
chains A list of rwty.chain objects
}
\description{
This function is automatically called by many other functions, but can be run manually as well.
It performs a number of tests of chain format, labels, lengths, etc.
}
\examples{
data(fungus)
check.chains(fungus)
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{rwty}
