% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.ptables.R
\name{combine.ptables}
\alias{combine.ptables}
\title{Function for merging p tables for multiple MCMC chains}
\usage{
combine.ptables(chains, burnin)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}
}
\value{
ptable A data frame of likelihood values and model parameters for the supplied rwty.trees objects
}
\description{
This function is automatically called by some of the plot functions.
}
\examples{
data(fungus)
combine.ptables(fungus, burnin=20)
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{rwty}

