% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.n.R
\name{compare.n}
\alias{compare.n}
\title{Compares posterior probability estimates from multiple MCMC chains.}
\usage{
compare.n(chains, burnin = 0, min.freq = 0)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}

\item{min.freq}{The minimum frequency for a node to be used for calculating Average Standard Deviation of Split Frequncies (ASDSF).  Default value is 0.0.}
}
\value{
output A list containing a table of frequencies of each clade in each chain along with mean and sd, a distance matrix measuring consensus between chains, a translation table, and a ggpairs plot.
}
\description{
This function takes a set of tree files, an option array of names for those
files, and a burnin argument.  It returns a table with the frequencies of
each clade in each of those tree files, the pairwise and grand mean average
and a translation table.
}
\examples{
data(fungus)
compare.n(fungus, burnin=20)
}
\keyword{MCMC,}
\keyword{clade}
\keyword{consensus,}
\keyword{convergence}
\keyword{frequency,}
\keyword{phylogenetics,}

