% Generated by roxygen2 (4.0.1): do not edit by hand
\name{continuous.distance}
\alias{continuous.distance}
\title{Various small utility functions}
\usage{
continuous.distance(two.trees)
}
\description{
A bunch of tiny utility functions that we need off and on.  Not intended for user interaction.
}

