% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cumulative.freq.R
\name{cumulative.freq}
\alias{cumulative.freq}
\title{Cumulative frequencies of clades in an MCMC chain}
\usage{
cumulative.freq(tree.list, burnin = 0, window.size, gens.per.tree = 1,
  slide.freq.table = NULL, ...)
}
\arguments{
\item{tree.list}{A single rwty.trees or multiPhylo object.}

\item{burnin}{The number of trees to eliminate as burnin.  Defaults to zero.}

\item{window.size}{The length of window (in trees) for the sliding window plot}

\item{gens.per.tree}{The number of generations per tree in the .t file.}

\item{slide.freq.table}{A table from slide.freq.  If a table is not provided, it will be built.}
}
\value{
output A list containing a table of cumulative frequencies and a clade translation table
}
\description{
Calculates the posterior probability estimate of each clade in an MCMC chain
as a function of chain length.
}
\examples{
data(fungus)
cumulative.freq(fungus$Fungus.Run1$trees, burnin=100, window.size=20, gens.per.tree=10000)
}
\keyword{MCMC,}
\keyword{convergence}
\keyword{phylogenetics,}
\keyword{posterior}
\keyword{probabilities,}

