% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.trees.R
\name{load.trees}
\alias{load.trees}
\title{Custom functions to load tree lists so that rwty can do basic processing on the way in.}
\usage{
load.trees(file, type = "nexus", gens.per.tree = NA, trim = 1,
  logfile = NA, skip = 1)
}
\arguments{
\item{file}{A path to a .t file containing an MCMC chain of trees}

\item{type}{An argument that designates the type of tree file.  If "nexus",
trees are loaded using ape's read.nexus function.  Otherwise, it's read.tree.}

\item{gens.per.tree}{The number of generations separating trees.  If not provided, RWTY will attempt to calculate it automatically.}

\item{trim}{Used for thinning the chain.  If a number N is provided, RWTY keeps every Nth tree.}

\item{logfile}{A path to a file containing model parameters and likelihoods}

\item{skip}{The number of lines that must be skipped to get to the header of the log file.
MrBayes, for instance, prints a comment line at the top of the log file, so MrBayes files should be
read in with a skip value of 1.}
}
\value{
output An rwty.trees object containing the multiPhylo and the table of values from the log file if available.
}
\description{
Loads trees, looks for a .p file of tree likelihoods, returns and rwty.trees object containing both
}
\examples{
#load.trees(file="mytrees.nex", type="nexus")
}
\keyword{MCMC,}
\keyword{Phylogenetics,}
\keyword{load}

