% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplot.autocorr.R
\name{makeplot.autocorr}
\alias{makeplot.autocorr}
\title{Make autocorrelation plots of tree topologies from MCMC analyses}
\usage{
makeplot.autocorr(
  chains,
  burnin = 0,
  max.sampling.interval = NA,
  autocorr.intervals = 40,
  facet = FALSE,
  free_y = FALSE
)
}
\arguments{
\item{chains}{A list of rwty.chain objects.}

\item{burnin}{The number of trees to eliminate as burnin.}

\item{max.sampling.interval}{The largest sampling interval for which you want to calculate the mean distance between pairs of trees (default is 10 percent of the length of the chain).}

\item{autocorr.intervals}{The number of sampling intervals to use. These will be spaced evenly between 1 and the max.sampling.interval}

\item{facet}{TRUE/FALSE to turn facetting of the plot on or off (default FALSE)}

\item{free_y}{TRUE/FALSE to turn free y scales on the facetted plots on or off (default FALSE). Only works if facet = TRUE.}
}
\value{
A ggplot2 plot object, with one line (facetting off) or facet
(facetting on) per rwty.chain object.
}
\description{
This function takes a list of rwty.chain objects, and makes an
autocorrelation plot for each chain. Each plot shows the mean phylogenetic
distance at a series of sampling intervals. In really well behaved MCMC 
analyses, the mean distance will stay constant as the sampling interval 
increases. If there is autocorrelation, the mean distance will 
increase as the sampling interval increases, and is expected to level
off when the autocorrelation decreases to zero. The function calculates
path distances, though other distances could also be employed.
}
\examples{
\dontrun{
data(fungus)
makeplot.autocorr(fungus, burnin = 20)
}
}
\keyword{autocorrelation,}
\keyword{distance}
\keyword{path}
