% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplot.autocorr.R
\name{makeplot.autocorr}
\alias{makeplot.autocorr}
\title{Make autocorrelation plots of tree topologies from MCMC analyses}
\usage{
makeplot.autocorr(chains, burnin = 0, max.intervals = 100,
  ac.cutoff = 0.95, facet = FALSE)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin.}

\item{max.intervals}{The maximum number of sampling intervals to use.}

\item{ac.cutoff}{The proportion of the estimated asymptotic path distance to use as a cutoff for estimating minimum sampling interval.  For instance, if ac.cutoff = 0.9, the sampling interval returned will be the minimum sampling interval necessary to achieve a path distance of 0.9 times the estimated asymptotic value.}

\item{facet}{TRUE/FALSE to turn facetting of the plot on or off (default)}
}
\value{
A ggplot2 plot object, with one line (facetting off) or facet
(facetting on) per rwty.trees object, and a data frame containing estimated minimum sampling intervals for each chain.
}
\description{
This function takes a list of rwty.trees objects, and makes an
autocorrelation plot for each chain. Each plot shows the median phylogenetic
distance at a series of sampling intervals. In really well behaved MCMC 
analyses, the median distance will stay constant as the sampling interval 
increases. If there is autocorrelation, the median distance will 
increase as the sampling interval increases, and is expected to level
off when the autocorrelation decreases to zero. The function calculates
path distances, though other distances could also be employed.
}
\examples{
data(fungus)
makeplot.autocorr(fungus, burnin = 20)
This bit uses an exponential semivariogram model to estimate the asymptotic path distance, and
uses that to estimate the sampling interval at which path distances have reached some fixed
proportion of that value (default 0.95).
}
\keyword{autocorrelation,}
\keyword{distance}
\keyword{path}

