% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplot.burnin.R
\name{makeplot.burnin}
\alias{makeplot.burnin}
\title{Plotting parameters}
\usage{
makeplot.burnin(chains, burnin = NA)
}
\arguments{
\item{chains}{A set of rwty.chain objects.}

\item{burnin}{The number of trees to omit as burnin. The default (NA) is to use the maximum burnin from all burnins calculated automatically when loading the chains. This can be overidden by providing any integer value.}
}
\value{
param.plot Returns a ggplot object.
}
\description{
Plots parameter values over the length of the MCMC chain
}
\examples{
\dontrun{
data(fungus)
makeplot.burnin(fungus)
}
}
\keyword{convergence,}
\keyword{mcmc,}
\keyword{parameter,}
\keyword{phylogenetics}
\keyword{plot,}
