% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.multichain.R
\name{makeplot.multichain}
\alias{makeplot.multichain}
\title{Function for running rwty analyses on multiple chains.}
\usage{
makeplot.multichain(chains, burnin, window.num, min.freq = 0.1)
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin.  Default is zero.}

\item{window.num}{The number of windows along the chain to use for analyses.}

\item{min.freq}{The minimum frequency for a node to be used for calculating ASDSF.}
}
\value{
output A list of plots illustrating agreement between chains.
}
\description{
This function is automatically called when analyze.simple is called with multple chains.
It runs analyze.single for each chain, and then does multi-chain analyses as well.
}
\examples{
data(fungus)
makeplot.multichain(fungus, burnin=20,  window.num = 20, min.freq = 0.1 )
}
\keyword{MCMC,}
\keyword{awty,}
\keyword{convergence,}
\keyword{phylogenetics,}
\keyword{plot,}
\keyword{rwty}

