% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeplot.treespace.R
\name{makeplot.treespace}
\alias{makeplot.treespace}
\title{Plot chains in treespace.}
\usage{
makeplot.treespace(chains, n.points = 100, burnin = 0, likelihood = NA)
}
\arguments{
\item{chains}{A list of one or more rwty.trees objects}

\item{n.points}{The number of points on each plot}

\item{burnin}{The number of samples to remove from the start of the chain as burnin}

\item{likelihood}{The name of the column that describes the likelihood in your parameter file, if present}
}
\value{
A list of two ggplot objects: one plots the points in treespace, the other shows a heatmap of the same points
}
\description{
This function will take an mds.treespace object and produce plots of chains in treespace.
}
\examples{
data(fungus)

p <- makeplot.treespace(fungus, burnin = 20, likelihood = 'LnL')

# NB: these data indicate significant problems: the two chains are sampling very different parts of tree space
# View the points plotted in treespace (these data indicate significant problems)
p$points.plot

# View the heatmap of the same data
p$heatmap
}
\keyword{plot,}
\keyword{rwty}
\keyword{treespace,}

