\name{plot.treespace}
\alias{plot.treespace}
\title{Plot chains in treespace.}
\usage{
  plot.treespace(points)
}
\arguments{
  \item{points}{A set of points to plot \code{points}}
}
\value{
  p A ggplot object
}
\description{
  This function will take an mds.treespace object and
  produce plots of chains in treespace.
}
\examples{
plot.treespace(mypoints)
}
\keyword{plot,}
\keyword{rwty}
\keyword{treespace,}

