% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.approx.ess.R
\name{topological.approx.ess}
\alias{topological.approx.ess}
\title{Calculate the approximate Effective Sample Size (ESS) of tree topologies}
\usage{
topological.approx.ess(chains, burnin = NA)
}
\arguments{
\item{chains}{A list of rwty.chain objects.}

\item{burnin}{The number of trees to omit as burnin. The default (NA) is to use the maximum burnin from all burnins calculated automatically when loading the chains. This can be overidden by providing any integer value.}
}
\value{
A data frame with one row per chain, and columns describing the
approximate ESS and the name of the chain.
}
\description{
This function takes a list of rwty.chain objects, and calculates the
approximate ESS of the trees from each chain, after removing burnin.
The calculation uses the autocorrelation among squared topological distances between
trees to calculate an approximate ESS of tree topologies for each chain.
NB: if all post-burnin trees are identical, the function will return an ESS value that equals the number of post burnin samples.
}
\examples{
\dontrun{
data(fungus)
topological.approx.ess(fungus, burnin = 20)
}
}
\keyword{distance}
\keyword{distance,}
\keyword{path}
\keyword{tree}
\keyword{treespace,}
