% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topological.approx.ess.R
\name{topological.approx.ess}
\alias{topological.approx.ess}
\title{Calculate the approximate Effective Sample Size (ESS) of tree topologies}
\usage{
topological.approx.ess(chains, burnin = 0, autocorr.intervals = NA,
  treedist = "PD")
}
\arguments{
\item{chains}{A list of rwty.trees objects.}

\item{burnin}{The number of trees to eliminate as burnin}

\item{autocorr.intervals}{The largest sampling interval you want to plot. Defaults to the largest possible sampling interval, which is the one for which we can sample at least 20 indpendent pairs of trees}

\item{treedist}{the type of tree distance metric to use, can be 'PD' for path distance or 'RF' for Robinson Foulds distance}
}
\value{
A data frame with one row per chain, and columns describing the
approximate ESS and the name of the chain.
}
\description{
This function takes a list of rwty.trees objects, and calculates the
pseudo ESS of the trees from each chain, after removing burnin.
The calculation uses the autocorrelation among squared topological distances between
trees to calculate an approximate ESS of tree topologies for each chain.
NB this function requires the calculation of many many
tree distances, so can take some time.
}
\examples{
data(fungus)
topological.approx.ess(fungus, burnin = 20)
}
\keyword{distance}
\keyword{distance,}
\keyword{path}
\keyword{tree}
\keyword{treespace,}

