% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.autocorr.R
\name{topological.autocorr}
\alias{topological.autocorr}
\title{Calculate data for autocorrelation plots of tree topologies from MCMC analyses}
\usage{
topological.autocorr(chains, burnin = NA, max.sampling.interval = NA)
}
\arguments{
\item{chains}{A list of rwty.chain objects.}

\item{burnin}{The number of trees to omit as burnin. The default (NA) is to use the maximum burnin from all burnins calculated automatically when loading the chains. This can be overidden by providing any integer value.}

\item{max.sampling.interval}{The largest sampling interval for which you want to calculate the mean distance between pairs of trees (default is the larger of 10 percent of the length of the chain, or the sampling interval that gives at least 100 paired samples).}
}
\value{
A data frame with one row per sampling interval, per chain. 
The first column is the sampling interval. The second column is the mean 
path distance between pairs of trees from that sampling interval. The third
column is the chain ID.
}
\description{
This function takes a list of rwty.chain objects, and calculates the
mean phylogenetic distance at a series of roughly even sampling intervals.
In really well behaved MCMC analyses, the mean distance will stay constant
as the sampling interval increases. If there is autocorrelation, it will 
increase as the sampling interval increases, and is expected to level
off when the autocorrelation decreases to zero. The function calculates
path distances, though other distances could also be employed.
}
\examples{
\dontrun{
data(fungus)
topological.autocorr(fungus, burnin = 20)
}
}
\keyword{autocorrelation,}
\keyword{distance}
\keyword{path}
