% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.dist.matrix.R
\name{tree.dist.matrix}
\alias{tree.dist.matrix}
\title{Tree distance matrix calculation}
\usage{
tree.dist.matrix(trees)
}
\arguments{
\item{trees}{a multiPhylo object}
}
\value{
RF A distance matrix of RF distances
}
\description{
This function takes a list of trees and returns a distance matrix
populated with Robinsoun-Foulds tree distances between all trees
in the list.
}
\examples{
data(fungus)
tree.dist.matrix(list(run1$trees[[1]], run1$trees[[2]], run1$trees[[3]]))
}
\keyword{distance,}
\keyword{robinson-foulds}
\keyword{tree}
\keyword{treespace,}

