% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.dist.matrix.R
\name{tree.dist.matrix}
\alias{tree.dist.matrix}
\title{Tree distance matrix calculation}
\usage{
tree.dist.matrix(trees, treedist = "RF")
}
\arguments{
\item{trees}{a multiPhylo object}

\item{treedist}{the type of tree distance metric to use, can be 'PD' for path distance or 'RF' (the default) for Robinson Foulds distance}
}
\value{
RF A distance matrix of RF distances
}
\description{
This function takes a list of trees and returns a distance matrix
populated with Robinson-Foulds tree distances between all trees
in the list.
}
\examples{
\dontrun{
data(fungus)
tree.dist.matrix(fungus$Fungus.Run1$trees)
}
}
\keyword{distance,}
\keyword{robinson-foulds}
\keyword{tree}
\keyword{treespace,}
