% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.dist.matrix.R
\name{tree.dist.matrix}
\alias{tree.dist.matrix}
\title{Tree distance matrix calculation}
\usage{
tree.dist.matrix(trees)
}
\arguments{
\item{trees}{a multiPhylo object}
}
\value{
RF A distance matrix of RF distances
}
\description{
This function takes a list of trees and returns a distance matrix
populated with Robinson-Foulds tree distances between all trees
in the list.
}
\examples{
data(fungus)
tree.dist.matrix(fungus$Fungus.Run1$trees)
}
\keyword{distance,}
\keyword{robinson-foulds}
\keyword{tree}
\keyword{treespace,}
