% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treespace.R
\name{treespace}
\alias{treespace}
\title{MDS scaling of treespace for a single tree list.}
\usage{
treespace(chains, n.points, burnin = 0, likelihood = NA)
}
\arguments{
\item{chains}{A list of 1 or more rwty.trees objects.}

\item{n.points}{The minimum number of points you want in your plot.}

\item{burnin}{The number of trees to eliminate as burnin. Default is zero.}
}
\value{
Returns a list containing the points and a plot.
}
\description{
This function constructs a distance matrix from a list of trees
and uses multi-dimensional scaling to collapse it to a two-
dimensional tree space for plotting.
}
\examples{
data(fungus)
treespace(fungus, n.points=50, burnin=100, likelihood="LnL")
}
\keyword{mds,}
\keyword{multi-dimensional}
\keyword{scaling}
\keyword{treespace,}

